fix: share app bundle icon persistence policy
