+++ Sources/AppIconDockTilePlugin.swift
@@ -4,28 +4,6 @@ import CoreServices
 private let cmuxAppIconDidChangeNotification = Notification.Name("com.cmuxterm.appIconDidChange")
 private let cmuxAppIconModeKey = "appIconMode"
 
-enum AppBundleIconPersistencePolicy {
-    private static let stableReleaseBundleIdentifier = "com.cmuxterm.app"
-    private static let stableReleaseAppBundleName = "cmux.app"
-    static let disablePersistenceArgument = "--cmux-disable-bundle-icon-persistence"
-
-    static func shouldPersist(
-        bundleIdentifier: String?,
-        appBundleLastPathComponent: String?,
-        launchArguments: [String] = ProcessInfo.processInfo.arguments
-    ) -> Bool {
-        guard !launchArguments.contains(disablePersistenceArgument) else {
-            return false
-        }
-
-        // Channel variants own their identity through build-time bundle metadata.
-        // Persisted Finder icons would override that metadata and can leak into
-        // packaged artifacts after CI smoke launches the app bundle.
-        return bundleIdentifier == stableReleaseBundleIdentifier
-            && appBundleLastPathComponent == stableReleaseAppBundleName
-    }
-}
-
 private enum DockTileAppIconMode: String {
     case automatic
     case light
